clean up code and move collection constants to constants.js
